Start to Encapsulate the search URL to image URL logic
This is still not perfect. It's serialized and just sort of hacky, but it is a start.
